% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{delta_create}
\alias{delta_create}
\title{Create a new empty Delta Lake table}
\usage{
delta_create(
  table_uri,
  schema,
  partition_by,
  name,
  description,
  storage_options,
  configuration
)
}
\arguments{
\item{table_uri}{Path where the table will be created}

\item{schema}{Arrow schema for the table}

\item{partition_by}{Column names to partition by (optional)}

\item{name}{Table name (optional)}

\item{description}{Table description (optional)}

\item{storage_options}{Storage backend options (optional)}

\item{configuration}{Table configuration properties (optional)}
}
\description{
Create a new empty Delta Lake table
}
