% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{delta_table_open}
\alias{delta_table_open}
\title{Open a Delta Table at the specified path}
\usage{
delta_table_open(path, storage_options)
}
\description{
Open a Delta Table at the specified path
}
