% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{get_files}
\alias{get_files}
\title{Get the list of Parquet files in the current table snapshot}
\usage{
get_files(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character vector of file URIs.
}
\description{
Returns the absolute URIs of all Parquet files that make up the current
version of the Delta table. These can be passed to other tools like
arrow, polars, or duckdb for reading.
}
\examples{
\dontrun{
dt <- delta_table("path/to/delta_table")
files <- get_files(dt)

# Use with arrow
arrow::open_dataset(files)
}

}
