% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{get_metadata}
\alias{get_metadata}
\title{Get table metadata}
\usage{
get_metadata(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}
}
\value{
A named list with table metadata.
}
\description{
Returns metadata about the Delta table including id, name, description,
partition columns, and configuration.
}
