% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{get_schema}
\alias{get_schema}
\title{Get table schema}
\usage{
get_schema(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An Arrow Schema object.
}
\description{
Returns the Arrow schema of the Delta table.
}
