% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{is_delta_table}
\alias{is_delta_table}
\title{Check if a path is a Delta Table}
\usage{
is_delta_table(path, storage_options)
}
\arguments{
\item{path}{Path to check.}

\item{storage_options}{Optional storage options for the backend.}
}
\description{
Check if a path is a Delta Table
}
