% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{is_delta_table_path}
\alias{is_delta_table_path}
\title{Check if a path contains a Delta table}
\usage{
is_delta_table_path(path, storage_options = NULL)
}
\arguments{
\item{path}{Character. Path to check.}

\item{storage_options}{Optional named list. Storage backend options.}
}
\value{
Logical. TRUE if the path contains a valid Delta table.
}
\description{
Check if a path contains a Delta table
}
