% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{load_version}
\alias{load_version}
\title{Load a specific version of the table}
\usage{
load_version(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}

\item{version}{Integer. The version number to load.}
}
\value{
The DeltaTable object (invisibly), updated to the specified version.
}
\description{
Updates the DeltaTable to point to a specific version.
}
