% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{partition_columns}
\alias{partition_columns}
\title{Get partition columns}
\usage{
partition_columns(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character vector of partition column names.
}
\description{
Returns the partition columns of the Delta table.
}
