% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{register_cloud_handlers}
\alias{register_cloud_handlers}
\title{Register cloud storage handlers (GCS, S3, Azure) for deltalake
Called from R's .onLoad to enable cloud storage support}
\usage{
register_cloud_handlers()
}
\description{
Register cloud storage handlers (GCS, S3, Azure) for deltalake
Called from R's .onLoad to enable cloud storage support
}
