% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_table.R
\name{table_version}
\alias{table_version}
\title{Get the current version of a Delta table}
\usage{
table_version(table, ...)
}
\arguments{
\item{table}{A DeltaTable object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Integer. The current version number.
}
\description{
Get the current version of a Delta table
}
